EntityPropertyTrait
